requirements.txt
apache-airflow
pandas
requests
beautifulsoup4
elasticsearch>=7.17,<8.0.0
psycopg2-binary
apache-airflow-providers-postgres
pyspark==3.5.0
pyarrow
